Group the telemetry files and rename the setup folder
