+++ Sources/Scout/Diagnostics/Telemetry/Bucket/LatencyBuckets.swift
