+++ Sources/Scout/Diagnostics/Telemetry/Bucket/StatusBuckets.swift
