+++ Sources/Scout/Diagnostics/Telemetry/Handler/TelemetryFactory.swift
