+++ Sources/Scout/Diagnostics/Telemetry/Handler/TelemetryHandler+LogMetrics.swift
