+++ Sources/Scout/Diagnostics/Telemetry/Model/Telemetry.Export.swift
