--- Sources/Scout/Diagnostics/Telemetry/ResetMarker.swift
@@ -1,11 +1,0 @@
-//
-// Copyright 2026 Mikhail Kasianov
-//
-// Use of this source code is governed by an MIT-style
-// license that can be found in the LICENSE file or at
-// https://opensource.org/licenses/MIT.
-//
-
-package enum ResetMarker {
-    package static let category = "counter_reset"
-}
