+++ Tests/ScoutTests/Diagnostics/Telemetry/Handler/TelemetryFactoryTests.swift
