+++ Tests/ScoutTests/Diagnostics/Telemetry/Handler/TelemetryHandlerTests.swift
